refactor: LottoMachine 도메인 추가
